Btrfs: set trans to null in reserve_metadata_bytes if we commit the transaction
Patchwork-id: 31632
O-Subject: [RHEL6.1 PATCH 10/81] Btrfs: set trans to null in
	reserve_metadata_bytes if we commit the transaction
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

btrfs_commit_transaction will free our trans, but because we pass trans to
shrink_delalloc we could possibly have a use after free situation.  So instead
if we commit the transaction, set trans to null and set committed to true so we
don't keep trying to commit a transaction.  This fixes a panic I could reproduce
at will.  Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>