Btrfs: load free space cache if it exists
Patchwork-id: 31637
O-Subject: [RHEL6.1 PATCH 14/81] Btrfs: load free space cache if it exists
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

This patch actually loads the free space cache if it exists.  The only thing
that really changes here is that we need to cache the block group if we're going
to remove an extent from it.  Previously we did not do this since the caching
kthread would pick it up.  With the on disk cache we don't have this luxury so
we need to make sure we read the on disk cache in first, and then remove the
extent, that way when the extent is unpinned the free space is added to the
block group.  This has been tested with all sorts of things.

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>