Btrfs: check cache->caching_ctl before returning if caching has started
Patchwork-id: 31638
O-Subject: [RHEL6.1 PATCH 15/81] Btrfs: check cache->caching_ctl before
	returning if caching has started
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

With the free space disk caching we can mark the block group as started with the
caching, but we don't have a caching ctl.  This can race with anybody else who
tries to get the caching ctl before we cache (this is very hard to do btw).  So
instead check to see if cache->caching_ctl is set, and if not return NULL.
Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>