Btrfs: Add a clear_cache mount option
Patchwork-id: 31639
O-Subject: [RHEL6.1 PATCH 17/81] Btrfs: Add a clear_cache mount option
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

If something goes wrong with the free space cache we need a way to make sure
it's not loaded on mount and that it's cleared for everybody.  When you pass the
clear_cache option it will make it so all block groups are setup to be cleared,
which keeps them from being loaded and then they will be truncated when the
transaction is committed.  Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>