Add new functions for triggering inode writeback
Patchwork-id: 31644
O-Subject: [RHEL6.1 PATCH 21/81] Add new functions for triggering inode writeback
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Chris Mason <chris.mason@oracle.com>

When btrfs is running low on metadata space, it needs to force delayed
allocation pages to disk.  It currently does this with a suboptimal walk
of a private list of inodes with delayed allocation, and it would be
much better if we used the generic flusher threads.

writeback_inodes_sb_if_idle would be ideal, but it waits for the flusher
thread to start IO on all the dirty pages in the FS before it returns.
This adds variants of writeback_inodes_sb* that allow the caller to
control how many pages get sent down.

Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>