Btrfs: fix raid code for removing missing drives
Patchwork-id: 31647
O-Subject: [RHEL6.1 PATCH 25/81] Btrfs: fix raid code for removing missing drives
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Chris Mason <chris.mason@oracle.com>

When btrfs is mounted in degraded mode, it has some internal structures
to track the missing devices.  This missing device is setup as readonly,
but the mapping code can get upset when we try to write to it.

This changes the mapping code to return -EIO instead of oops when we try
to write to the readonly device.

Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>