Btrfs: use dget_parent where we can UPDATED
Patchwork-id: 31661
O-Subject: [RHEL6.1 PATCH 54/81] Btrfs: use dget_parent where we can UPDATED
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

There are lots of places where we do dentry->d_parent->d_inode without holding
the dentry->d_lock.  This could cause problems with rename.  So instead we need
to use dget_parent() and hold the reference to the parent as long as we are
going to use it's inode and then dput it at the end.

Signed-off-by: Josef Bacik <josef@redhat.com>
Cc: raven@themaw.net
Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>