Btrfs: hold i_mutex when calling btrfs_log_dentry_safe
Patchwork-id: 31696
O-Subject: [RHEL6.1 PATCH 55/81] Btrfs: hold i_mutex when calling
	btrfs_log_dentry_safe
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Since we walk up the path logging all of the parts of the inode's path, we need
to hold i_mutex to make sure that the inode is not renamed while we're logging
everything.  btrfs_log_dentry_safe does dget_parent and all of that jazz, but we
may get unexpected results if the rename changes the inode's location while
we're higher up the path logging those dentries, so do this for safety reasons.
Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>