Btrfs: make sure new inode size is ok in fallocate
Patchwork-id: 31697
O-Subject: [RHEL6.1 PATCH 59/81] Btrfs: make sure new inode size is ok in
	fallocate
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

We have been failing xfstest 228 forever, because we don't check to make sure
the new inode size is acceptable as far as RLIMIT is concerned.  Just check to
make sure it's ok to create a inode with this new size and error out if not.
With this patch we now pass 228.

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Chris Mason <chris.mason@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>