Btrfs: do not do fast caching if we are allocating blocks for tree_root
Patchwork-id: 31686
O-Subject: [RHEL6.1 PATCH 68/81] Btrfs: do not do fast caching if we are
	allocating blocks for tree_root
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Since the fast caching uses normal tree locking, we can possibly deadlock if we
get to the caching via a btrfs_search_slot() on the tree_root.  So just check to
see if the root we are on is the tree root, and just don't do the fast caching.

Reported-by: Sage Weil <sage@newdream.net>
Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>