Btrfs: do not BUG if we fail to remove the orphan item for dead snapshots
Patchwork-id: 31700
O-Subject: [RHEL6.1 PATCH 70/81] Btrfs: do not BUG if we fail to remove the
	orphan item for dead snapshots
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Not being able to delete an orphan item isn't a horrible thing.  The worst that
happens is the next time around we try and do the orphan cleanup and we can't
find the referenced object and just delete the item and move on.

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>