md/raid5: ensure we create a unique name for kmem_cache when mddev has no gendisk
Patchwork-id: 31742
O-Subject: [RHEL6.1 Patch 33/80] md/raid5: ensure we create a unique name for
	kmem_cache when mddev has no gendisk
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

We will shortly allow md devices with no gendisk (they are attached to
a dm-target instead).  That will cause mdname() to return 'mdX'.
There is one place where mdname really needs to be unique: when
creating the name for a slab cache.
So in that case, if there is no gendisk, you the address of the mddev
formatted in HEX to provide a unique name.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>