md/plug: optionally use plugger to unplug an array during resync/recovery.
Patchwork-id: 31777
O-Subject: [RHEL6.1 Patch 41/80] md/plug: optionally use plugger to unplug an
	array during resync/recovery.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

If an array doesn't have a 'queue' then md_do_sync cannot
unplug it.
In that case it will have a 'plugger', so make that available
to the mddev, and use it to unplug the array if needed.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>