md/bitmap: prepare for storing write-intent-bitmap via dm-dirty-log.
Patchwork-id: 31779
O-Subject: [RHEL6.1 Patch 47/80] md/bitmap: prepare for storing
	write-intent-bitmap via dm-dirty-log.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

This allows md/raid5 to fully work as a dm target.

Normally md uses a 'filemap' which contains a list of pages of bits
each of which may be written separately.
dm-log uses and all-or-nothing approach to writing the log, so
when using a dm-log, ->filemap is NULL and the flags normally stored
in filemap_attr are stored in ->logattrs instead.

Signed-off-by: NeilBrown <neilb@suse.de>