Update recovery_offset even when external metadata is used.
Patchwork-id: 31783
O-Subject: [RHEL6.1 Patch 54/80] Update recovery_offset even when external
	metadata is used.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

The update of ->recovery_offset in sync_sbs is appropriate even then external
metadata is in use.  However sync_sbs is only called when native
metadata is used.

So move that update in to the top of md_update_sb (which is the only
caller of sync_sbs) before the test on ->external.

This moves the update out of ->write_lock protection, but those fields
only need ->reconfig_mutex protection which they still have.

Also move the test on ->persistent up to where ->external is set as
for metadata update purposes they are the same.

Clear MD_CHANGE_DEVS and MD_CHANGE_CLEAN as they can only be confusing
if ->external is set or ->persistent isn't.

Finally move the update of ->utime down as it is only relevent (like
the ->events update) for native metadata.

Signed-off-by: NeilBrown <neilb@suse.de>
Reported-by: "Kwolek, Adam" <adam.kwolek@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>