md: Notify sysfs when RAID1/5/10 disk is In_sync.
Patchwork-id: 31776
O-Subject: [RHEL6.1 Patch 55/80] md: Notify sysfs when RAID1/5/10 disk is
	In_sync.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: Adrian Drzewiecki <adriand@vmware.com>

When RAID1 is done syncing disks, it'll update the state
of synced rdevs to In_sync. But it neglected to notify
sysfs that the attribute changed. So any programs that
are waiting for an rdev's state to change will not be
woken.

(raid5/raid10 added by neilb)

Signed-off-by: Adrian Drzewiecki <adriand@vmware.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>