md: call md_update_sb even for 'external' metadata arrays.
Patchwork-id: 31778
O-Subject: [RHEL6.1 Patch 59/80] md: call md_update_sb even for 'external'
	metadata arrays.
Bugzilla: 633695 659623
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

Now that we depend on md_update_sb to clear variable bits in
mddev->flags (rather than trying not to set them) it is important to
always call md_update_sb when appropriate.

md_check_recovery has this job but explicitly avoids it for ->external
metadata arrays.  This is not longer appropraite, or needed.

However we do want to avoid taking the mddev lock if only
MD_CHANGE_PENDING is set as that is not cleared by md_update_sb for
external-metadata arrays.

Reported-by:  "Kwolek, Adam" <adam.kwolek@intel.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>